cors issues fixed 8

+++ app.apimstec.com/routes/api.php
@@ -62,7 +62,7 @@ Route::middleware(['web', 'auth', 'verified', 'active.domain'])->group(function 
         Route::get('/{page}/seo', [PageController::class, 'seo'])->name('seo');
         Route::put('/{page}', [PageController::class, 'update'])->name('update');
         Route::put('/{page}/seo', [PageController::class, 'updateSeo'])->name('update-seo');
-        Route::delete('/{page}', [PageController::class, 'destroy'])->name('destroy');
+        Route::delete('/{page}', [PageController::class, 'destroy'])->name('destroy')->middleware('permission:content.delete');
         Route::patch('/{page}/status', [PageController::class, 'updateStatus'])->name('update-status');
         Route::post('/{page}/toggle-publish', [PageController::class, 'togglePublish'])->name('toggle-publish');
     });
@@ -73,7 +73,7 @@ Route::middleware(['web', 'auth', 'verified', 'active.domain'])->group(function 
         Route::post('/', [BlogController::class, 'store'])->name('store');
         Route::get('/{blog}/edit', [BlogController::class, 'edit'])->name('edit');
         Route::put('/{blog}', [BlogController::class, 'update'])->name('update');
-        Route::delete('/{blog}', [BlogController::class, 'destroy'])->name('destroy');
+        Route::delete('/{blog}', [BlogController::class, 'destroy'])->name('destroy')->middleware('permission:content.delete');
         Route::patch('/{blog}/status', [BlogController::class, 'updateStatus'])->name('update-status');
         Route::post('/{blog}/toggle-publish', [BlogController::class, 'togglePublish'])->name('toggle-publish');
         Route::patch('/{blog}/visibility', [BlogController::class, 'updateVisibility'])->name('update-visibility');
